refactor: rename nearbyGPSCoordinate filename
