+++ src/modules/location/nearby-gps-coordinate.ts
